+++ flake.nix
@@ -17,18 +17,19 @@
       );
       hsSrc = root: with lib.fileset; toSource {
         inherit root;
-        fileset = fileFilter (file: any file.hasExt ["cabal" "hs" "md"] || file.type == "directory") ./.;
+        fileset = fileFilter (file: any file.hasExt [ "cabal" "hs" "md" ] || file.type == "directory") ./.;
       };
       pname = "lsp-client";
       src = hsSrc ./.;
       ghcs = [ "ghc92" "ghc94" "ghc96" "ghc98" ];
       overlay = final: prev: {
         haskell = prev.haskell // {
-          packageOverrides = lib.composeExtensions
+          packageOverrides = lib.composeManyExtensions [
             prev.haskell.packageOverrides
             (hfinal: hprev: {
-              "${pname}" = hfinal.callCabal2nix pname src { };
-            });
+              ${pname} = hfinal.callCabal2nix pname src { };
+            })
+          ];
         };
       };
     in
@@ -58,7 +59,7 @@
         in
         {
           formatter.${system} = pkgs.nixpkgs-fmt;
-          legacyPackages.${system} = { inherit (pkgs) haskell haskellPackages; };
+          legacyPackages.${system} = pkgs'.extend lspOverlay;
           packages.${system} = { inherit default; };
           devShells.${system} =
             foreach hps (ghcName: hp: {